+++ CHANGELOG.md
@@ -31,7 +31,7 @@ from contracts import contract
 ```
 ## 3.3.5
 
-- add `check_route` to `route_manhattan` to ensure correct route connectivity
+- `route_manhattan` ensures correct route connectivity
 
 ## 3.3.4